Refactor Middleware tests to separate files
