--- tests/MiddlewareTest.php
@@ -1,424 +1,0 @@
-<?php
-
-namespace Spatie\Permission\Test;
-
-use Illuminate\Http\Request;
-use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Auth;
-use InvalidArgumentException;
-use Spatie\Permission\Contracts\Permission;
-use Spatie\Permission\Exceptions\UnauthorizedException;
-use Spatie\Permission\Middlewares\PermissionMiddleware;
-use Spatie\Permission\Middlewares\RoleMiddleware;
-use Spatie\Permission\Middlewares\RoleOrPermissionMiddleware;
-
-class MiddlewareTest extends TestCase
-{
-    protected $roleMiddleware;
-    protected $permissionMiddleware;
-    protected $roleOrPermissionMiddleware;
-
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->roleMiddleware = new RoleMiddleware();
-
-        $this->permissionMiddleware = new PermissionMiddleware();
-
-        $this->roleOrPermissionMiddleware = new RoleOrPermissionMiddleware();
-    }
-
-    /** @test */
-    public function a_guest_cannot_access_a_route_protected_by_the_role_or_permission_middleware()
-    {
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'testRole')
-        );
-    }
-
-    /** @test */
-    public function a_guest_cannot_access_a_route_protected_by_rolemiddleware()
-    {
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, 'testRole')
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_role_middleware_of_another_guard()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, 'testAdminRole')
-        );
-    }
-
-    /** @test */
-    public function a_user_can_access_a_route_protected_by_role_middleware_if_have_this_role()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleMiddleware, 'testRole')
-        );
-    }
-
-    /** @test */
-    public function a_user_can_access_a_route_protected_by_this_role_middleware_if_have_one_of_the_roles()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleMiddleware, 'testRole|testRole2')
-        );
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleMiddleware, ['testRole2', 'testRole'])
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_the_role_middleware_if_have_a_different_role()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole(['testRole']);
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, 'testRole2')
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_role_middleware_if_have_not_roles()
-    {
-        Auth::login($this->testUser);
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, 'testRole|testRole2')
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_role_middleware_if_role_is_undefined()
-    {
-        Auth::login($this->testUser);
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, '')
-        );
-    }
-
-    /** @test */
-    public function a_guest_cannot_access_a_route_protected_by_the_permission_middleware()
-    {
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles')
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_the_permission_middleware_of_a_different_guard()
-    {
-        // These permissions are created fresh here in reverse order of guard being applied, so they are not "found first" in the db lookup when matching
-        app(Permission::class)->create(['name' => 'admin-permission2', 'guard_name' => 'web']);
-        app(Permission::class)->create(['name' => 'admin-permission2', 'guard_name' => 'admin']);
-        app(Permission::class)->create(['name' => 'edit-articles2', 'guard_name' => 'admin']);
-        app(Permission::class)->create(['name' => 'edit-articles2', 'guard_name' => 'web']);
-
-        Auth::login($this->testAdmin);
-
-        $this->testAdmin->givePermissionTo('admin-permission2');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, 'admin-permission2')
-        );
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles2')
-        );
-
-        Auth::login($this->testUser);
-
-        $this->testUser->givePermissionTo('edit-articles2');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles2')
-        );
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'admin-permission2')
-        );
-    }
-
-    /** @test */
-    public function a_user_can_access_a_route_protected_by_permission_middleware_if_have_this_permission()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles')
-        );
-    }
-
-    /** @test */
-    public function a_user_can_access_a_route_protected_by_this_permission_middleware_if_have_one_of_the_permissions()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-news|edit-articles')
-        );
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, ['edit-news', 'edit-articles'])
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_the_permission_middleware_if_have_a_different_permission()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-news')
-        );
-    }
-
-    /** @test */
-    public function a_user_cannot_access_a_route_protected_by_permission_middleware_if_have_not_permissions()
-    {
-        Auth::login($this->testUser);
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles|edit-news')
-        );
-    }
-
-    /** @test */
-    public function a_user_can_access_a_route_protected_by_permission_or_role_middleware_if_has_this_permission_or_role()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'testRole|edit-news|edit-articles')
-        );
-
-        $this->testUser->removeRole('testRole');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'testRole|edit-articles')
-        );
-
-        $this->testUser->revokePermissionTo('edit-articles');
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'testRole|edit-articles')
-        );
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, ['testRole', 'edit-articles'])
-        );
-    }
-
-    /** @test */
-    public function a_user_can_not_access_a_route_protected_by_permission_or_role_middleware_if_have_not_this_permission_and_role()
-    {
-        Auth::login($this->testUser);
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'testRole|edit-articles')
-        );
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'missingRole|missingPermission')
-        );
-    }
-
-    /** @test */
-    public function the_required_roles_can_be_fetched_from_the_exception()
-    {
-        Auth::login($this->testUser);
-
-        $requiredRoles = [];
-
-        try {
-            $this->roleMiddleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, 'some-role');
-        } catch (UnauthorizedException $e) {
-            $requiredRoles = $e->getRequiredRoles();
-        }
-
-        $this->assertEquals(['some-role'], $requiredRoles);
-    }
-
-    /** @test */
-    public function the_required_permissions_can_be_fetched_from_the_exception()
-    {
-        Auth::login($this->testUser);
-
-        $requiredPermissions = [];
-
-        try {
-            $this->permissionMiddleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, 'some-permission');
-        } catch (UnauthorizedException $e) {
-            $requiredPermissions = $e->getRequiredPermissions();
-        }
-
-        $this->assertEquals(['some-permission'], $requiredPermissions);
-    }
-
-    /** @test */
-    public function use_not_existing_custom_guard_in_role_or_permission()
-    {
-        $class = null;
-
-        try {
-            $this->roleOrPermissionMiddleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, 'testRole', 'xxx');
-        } catch (InvalidArgumentException $e) {
-            $class = get_class($e);
-        }
-
-        $this->assertEquals(InvalidArgumentException::class, $class);
-    }
-
-    /** @test */
-    public function use_not_existing_custom_guard_in_permission()
-    {
-        $class = null;
-
-        try {
-            $this->permissionMiddleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, 'edit-articles', 'xxx');
-        } catch (InvalidArgumentException $e) {
-            $class = get_class($e);
-        }
-
-        $this->assertEquals(InvalidArgumentException::class, $class);
-    }
-
-    /** @test */
-    public function use_not_existing_custom_guard_in_role()
-    {
-        $class = null;
-
-        try {
-            $this->roleMiddleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, 'testRole', 'xxx');
-        } catch (InvalidArgumentException $e) {
-            $class = get_class($e);
-        }
-
-        $this->assertEquals(InvalidArgumentException::class, $class);
-    }
-
-    /** @test */
-    public function user_can_not_access_role_with_guard_admin_while_using_default_guard()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleMiddleware, 'testRole', 'admin')
-        );
-    }
-
-    /** @test */
-    public function user_can_access_role_with_guard_admin_while_using_default_guard()
-    {
-        Auth::guard('admin')->login($this->testAdmin);
-
-        $this->testAdmin->assignRole('testAdminRole');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleMiddleware, 'testAdminRole', 'admin')
-        );
-    }
-
-    /** @test */
-    public function user_can_not_access_permission_with_guard_admin_while_using_default_guard()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->permissionMiddleware, 'edit-articles', 'admin')
-        );
-    }
-
-    /** @test */
-    public function user_can_access_permission_with_guard_admin_while_using_default_guard()
-    {
-        Auth::guard('admin')->login($this->testAdmin);
-
-        $this->testAdmin->givePermissionTo('admin-permission');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->permissionMiddleware, 'admin-permission', 'admin')
-        );
-    }
-
-    /** @test */
-    public function user_can_not_access_permission_or_role_with_guard_admin_while_using_default_guard()
-    {
-        Auth::login($this->testUser);
-
-        $this->testUser->assignRole('testRole');
-        $this->testUser->givePermissionTo('edit-articles');
-
-        $this->assertEquals(403,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'edit-articles|testRole', 'admin')
-        );
-    }
-
-    /** @test */
-    public function user_can_access_permission_or_role_with_guard_admin_while_using_default_guard()
-    {
-        Auth::guard('admin')->login($this->testAdmin);
-
-        $this->testAdmin->assignRole('testAdminRole');
-        $this->testAdmin->givePermissionTo('admin-permission');
-
-        $this->assertEquals(200,
-            $this->runMiddleware($this->roleOrPermissionMiddleware, 'admin-permission|testAdminRole', 'admin')
-        );
-    }
-
-    protected function runMiddleware($middleware, $parameter, $guard = null)
-    {
-        try {
-            return $middleware->handle(new Request(), function () {
-                return (new Response())->setContent('<html></html>');
-            }, $parameter, $guard)->status();
-        } catch (UnauthorizedException $e) {
-            return $e->getStatusCode();
-        }
-    }
-}
